+++ toupy/restoration/__init__.py
@@ -9,3 +9,4 @@ from .ramptools import *
 from .derivativetools import *
 from .phase_retrieval import *
 from .ring_correction import *
+from .holo_ctf import *